clear warnings
git-svn-id: svn://orca.physics.unc.edu/Orca64@11923 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Main Dialogs/CommandCenter/Scripting/ORScriptRunner.m
@@ -39,9 +39,9 @@ extern long num_lines;
 extern id functionList;
 extern int OrcaScriptparse(void);
 ORScriptRunner* theScriptRunner = nil;
-int OrcaScriptYYINPUT(char* theBuffer,int maxSize) 
+int OrcaScriptYYINPUT(char* theBuffer,unsigned long maxSize) 
 {
-	return (int)[theScriptRunner yyinputToBuffer:theBuffer withSize:maxSize];
+	return (int)[theScriptRunner yyinputToBuffer:theBuffer withSize:(int)maxSize];
 }
 //========================================================================
